requirements.txt
tensorflow==2.15.1
streamlit
numpy
pillow
streamlit-drawable-canvas
scipy